Separated the express code from app.js
express code can be seen at express.js
